+++ src/logo/builtin.c
@@ -1561,6 +1561,16 @@ static const FFlogo E[] = {
             FF_COLOR_FG_DEFAULT,
         },
     },
+    // EN-OS
+    {
+        .names = {"ENOS"},
+        .lines = FASTFETCH_DATATEXT_LOGO_ENOS,
+        .colors = {
+            FF_COLOR_FG_LIGHT_BLUE,
+        },
+        .colorKeys = FF_COLOR_FG_LIGHT_BLUE,
+        .colorTitle = FF_COLOR_FG_LIGHT_BLUE,
+    },
     // EncryptOS
     {
         .names = { "EncryptOS" },
@@ -2652,6 +2662,16 @@ static const FFlogo L[] = {
         },
         .colorKeys = FF_COLOR_FG_BLUE,
         .colorTitle = FF_COLOR_FG_BLUE,
+    },
+    // LimeOS
+    {
+        .names = {"LimeOS"},
+        .lines = FASTFETCH_DATATEXT_LOGO_LIMEOS,
+        .colors = {
+            FF_COLOR_FG_DEFAULT,
+        },
+        .colorKeys = FF_COLOR_FG_DEFAULT,
+        .colorTitle = FF_COLOR_FG_DEFAULT,
     },
     // Linspire
     {
@@ -4242,6 +4262,17 @@ static const FFlogo R[] = {
         .colorKeys = FF_COLOR_FG_RED,
         .colorTitle = FF_COLOR_FG_RED,
     },
+    // Redrose
+    {
+        .names = {"Redrose"},
+        .lines = FASTFETCH_DATATEXT_LOGO_REDROSE,
+        .colors = {
+            FF_COLOR_FG_RED,
+            FF_COLOR_FG_GREEN,
+        },
+        .colorKeys = FF_COLOR_FG_RED,
+        .colorTitle = FF_COLOR_FG_RED,
+    },
     // Refracta
     {
         .names = { "Refracta" },
@@ -5133,6 +5164,13 @@ static const FFlogo U[] = {
         },
     },
     // LAST
+    {
+        .names = { "Uzbek" },
+        .lines = FASTFETCH_DATATEXT_LOGO_UZBEK,
+        .colors = {
+            FF_COLOR_FG_GREEN,
+        },
+    },
     {},
 };